refactor: useState hook 규칙에 맞게 tech stack 관련 컴포넌트 리팩토링
